changed custom geolocation tool location
