+++ server/config/application.rb
@@ -5,6 +5,7 @@ require "./lib/monitor/monitor.rb"
 require "./lib/geotools/asns.rb"
 require "./lib/events_notifier/notifier.rb"
 require "./lib/tailscale/client.rb"
+require "./lib/geotools/custom_geocoder.rb"
 
 # Require the gems listed in Gemfile, including any gems
 # you've limited to :test, :development, or :production.
